requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-datastore==1.15.0
google-cloud-ndb==1.6.0
google-cloud-tasks==2.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
html2text==2020.1.16
httplib2==0.18.1
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
libcst==0.3.10
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.0
proto-plus==1.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pymemcache==3.4.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sendgrid==6.4.6
six==1.15.0
starkbank-ecdsa==1.1.0
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
